requirements/base.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.2
